Remove root_path from commands that read index
The root_path can be read from the headers of the index file, so there's
no point in including the former as an argument.

If the root_path has changed, that's a problem, but that will soon be
solved by a new option to update(), which will modify the root_path
header.