Include ignored directories in the ignored output
If we just prune the ignored directories without mentioning them, we're
clearly lying. We don't want to include each file in the tree under an
ignored dir, though, as that's just noise; what matters is which thing
was ignored.

So this adds the path to each ignored subdirectory to the ignored list.