Replace Index.headers with actual ivars
Writing index.headers['root_path'] has always felt wasteful. Since the
file format is versioned, we can have expectations about the presence of
required headers like 'root_path'. Adding new required headers can just
bump the version number.

The read() classmethod has also been reintroduced. This is because
__init__() was inelegantly accepting either a stream or a root_path, but
not both. A root_path would always be required. So instead of an ugly
bifurcated init, it's cleaner to have an init that demands root_path and
simply assigns ivars, and have a factory class method that gets the
required headers out of the file and hands them to the init.