Make Index.read() accept and open a path
This changes the index_file param from an open filehandle to a path.

We want to be able to look for a default index file name, and to look
for it in all parent dirs (i.e. search from cwd up, looking for
'.mnindex').

In order to do that, we need to accept a path, instead of letting click
deal with file handles for us.