Add 'date' header to all test index files
Doing this now will allow all tests to pass once we start transitioning
to the write-new-index renaming method of dealing with index files. This
is because in the new code, the presence of the 'date' header is
assumed, since it has always been included in v1 of the file format.

(This highlights the fact that I should write up a short spec of the
file formats, to make explicit what is included.)